fix float/int crash

+++ src/daemon/client.py
@@ -1434,7 +1434,7 @@ class Client(ServerSender, ray.ClientData):
                 _translate('GUIMSG', '%s seems to can not open')
                     % self.gui_msg_style())
 
-        duration = max(8000, 2 * self.last_open_duration)
+        duration = max(8000, int(2 * self.last_open_duration))
         self._open_timer.setInterval(duration)
         self._open_timer.start()
